+++ src/test/java/ui/runner/ApiUtils.java
@@ -1,20 +1,19 @@
-package ui.model.api;
+package ui.runner;
 
 import io.restassured.response.Response;
 import org.openqa.selenium.WebDriver;
 import ui.model.base.BasePage;
-import ui.runner.ProjectUtils;
 
 import static io.restassured.RestAssured.given;
 
-public class ApiPage extends BasePage {
+public class ApiUtils extends BasePage {
 
     private static final String apiLogin = ProjectUtils.getAdminName();
     private static final String apiPassword = ProjectUtils.getAdminPassword();
     private static final String apiUrl = "wp-json/wp/v2";
     private static final String USERS_LIST = "/users?context=view&page=1&per_page=10&order=desc&orderby=id";
 
-    public ApiPage(WebDriver driver) {
+    public ApiUtils(WebDriver driver) {
         super(driver);
     }
